+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # seyren cookbook CHANGELOG
 This file is used to list changes made in each version of the seyren cookbook.
 
+## v0.3.0 (2014-03-24)
+- Add seyren::mongo-import-data recipe.
+
 ## v0.2.0 (2014-03-24)
-- Change war uri to bintray. Chef have some probleme with github's redirect.
+- Change war uri to bintray. Chef has some probleme with github's redirect.
 
 ## v0.1.0 (2014-03-19)
-- Initial release of seyren
+- Initial release of seyren.